+++ lib/common.dart
@@ -12,5 +12,6 @@ export 'src/common.dart' show
     create_canister,
     top_up_canister,
     check_canister_status,
-    put_code_on_the_canister
+    put_code_on_the_canister,
+    IcpTokens
     ;